+++ src/Krystal/Date/CHANGELOG.md
@@ -4,6 +4,7 @@ CHANGELOG
 1.3
 ---
 
+ * In `TimeHelper` added `isTimestamp()` to test whether given string is UNIX-timestamp
  * In `TimeHelper` added `formatLocalized()` to format dates with localization support
  * In `TimeHelper` added `guessSeason()` to determine current season
  * In `TimeHelper` added `formatValid()` methods that checks whether date/time format is valid